fix: handle audit log with no entries
If a decision is invoked with a non-existing decision id, the engine returns a failure with an empty audit log. This caused an exception because the audit log can't handle an empty list.

Remove the public properties `rootEntry` and `requiredEntries` from the audit log. These properties doesn't work if the audit log entries are empty.

Note that the property `rootEntry` may contained an audit log entry but no from the root decision. If the evaluation of a required decision failed, the engine doesn't add an entry for the root decision.