feat: Handle null as unary-tests result
With the new version of the FEEL engine, a unary-tests expression may return `null`. For example, if a variable doesn't exist or if the value types are not comparable. Previously, the FEEL engine failed the evaluation or returned false.

To adopt the new behavior, treat `null` as `false`. In this case, the unary-tests expression was not fulfilled and the rule doesn't match.