test: Adjust expected behavior
With the new version of the FEEL engine, a non-existing variable results in `null`. Previously, the evaluation failed with an error. As a result, the evaluation of these test cases don't fail anymore.

To keep the semantics of the test case, force the UNIQUE hit policy to fail because of multiple matching rules.